Add quotes to UPDATE_SOURCESETS
At least on Linux/bash the quotes are necessary. Otherwise it is interpreted as a separate command.
Not really important, more of a convenience fix

Also, thank you for creating this helpful remapper tool! :)
If you prefer it, feel free to manually add the quote. PR was just the easiest way for me